Flattened some unnecessary presentation logic
In the team progress view there were some loops over lists provided
by the route for the purpose of displaying headers. I removed these and
replaced them with some static HTML and an extra argument for the
write_table view function, decreasing complexity and coupling between
the route and view.